+++ src/model/client/AbstractClient.types.ts
@@ -1,3 +1,4 @@
+import {type Interceptors} from './middleware/Interceptors';
 import {type RequestHeaders, type ResponseWrapper} from '@/types/request.types';
 import {type AbstractEndpoint} from '@/model/endpoint/AbstractEndpoint';
 import {type AbstractClient} from '@/model/client/AbstractClient';
@@ -19,6 +20,10 @@ export interface ClientConfig {
   parameters?: Record<string, string>;
   plugins?: PluginConstructor[];
   options?: ClientOptions;
+  interceptors?: {
+    request: Interceptors<RequestInit>;
+    response: Interceptors<Response>;
+  };
 }
 
 export interface ClientOptions {